using curried function
